// ui/src/generated/AudioToMidiCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AudioToMidiMode } from "./AudioToMidiMode";
import type { AudioToMidiOptions } from "./AudioToMidiOptions";
import type { ClipId } from "./ClipId";
import type { DeviceId } from "./DeviceId";
import type { NoteId } from "./NoteId";
import type { TrackId } from "./TrackId";

export type AudioToMidiCommand = { "type": "Start", job: string, clip: ClipId, mode: AudioToMidiMode, options: AudioToMidiOptions, track: TrackId, new_clip: ClipId, seed_notes: NoteId, instrument: DeviceId | null, } | { "type": "Cancel", job: string, };